xcp_debug_tool/requirements.txt
PyQt5>=5.15.0
pyelftools>=0.29
pywin32>=306
pytest>=7.0.0
